+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.4] 31-07-2023
+- Aplicamos correccion al test:e2e de div por 0
+- Agregamos test:e2e para multiplicacion con primer valor negativo
+- Agregamos funcionalidad a la suma
+- Agregamos test:e2e para la suma
+- Agregamos la funcion de limpieza de entrada con la tecla 'c'
+
 
 ## [1.3.3]